+++ cmd/seeder/v2/2_users.go
@@ -3,10 +3,11 @@ package main
 import (
 	"context"
 	"fmt"
+
 	"github.com/volatiletech/sqlboiler/queries/qm"
 	seeders "github.com/wednesday-solutions/go-template/cmd/seeder"
+	"github.com/wednesday-solutions/go-template/internal/postgres"
 	"github.com/wednesday-solutions/go-template/models"
-	"github.com/wednesday-solutions/go-template/pkg/utl/postgres"
 	"github.com/wednesday-solutions/go-template/pkg/utl/secure"
 )
