+++ internal/config/testdata/config.invalid.yaml
